Bot - Project/requirements.txt
sentence-transformers
faiss-cpu
pdfplumber
streamlit
streamlit-authenticator
numpy
tqdm
python-dateutil
scikit-learn